Update nvidia-smi features directory
Background
==========
A change was introduced in the latest version of the NVIDIA Grid Drivers which breaks
the install-gpu-drivers.ps1 script. Lines 113-121 in the script fail because the nvidia-smi
features directory has changed with the latest driver versions.
When installing the latest GRID /NVIDIA driver, the  nvidia-smi.exe is in the
C:\Windows\System32\DriverStore\FileRepository\nvgrid*\ instead of  C:\Program Files\NVIDIA Corporation\NVSMI

Changes
=======
- Update the latest GRID /NVIDIA driver location to C:\Windows\System32\DriverStore\FileRepository\nvgrid*\ instead of  C:\Program Files\NVIDIA Corporation\NVSMI
- update nodejs and python version from buildspec.yml